refactor(grid): Rename barrel file
