+++ projects/igniteui-angular/src/lib/grids/grid/expandable-cell.component.ts
@@ -1,10 +1,20 @@
-import { Component, ChangeDetectorRef, ElementRef, ViewChild, Inject,
-     ChangeDetectionStrategy, NgZone, OnInit, Input, TemplateRef } from '@angular/core';
+import {
+    ChangeDetectionStrategy,
+    ChangeDetectorRef,
+    Component,
+    ElementRef,
+    Inject,
+    Input,
+    NgZone,
+    OnInit,
+    TemplateRef,
+    ViewChild
+} from '@angular/core';
 import { IgxGridCellComponent } from '../cell.component';
 import { GridBaseAPIService } from '../api.service';
 import { getNodeSizeViaRange, PlatformUtil } from '../../core/utils';
 import { DOCUMENT } from '@angular/common';
-import { IgxGridBaseDirective } from '.';
+import { IgxGridBaseDirective } from './public_api';
 import { IgxGridSelectionService, IgxGridCRUDService } from '../selection/selection.service';
 import { HammerGesturesManager } from '../../core/touch';
 import { GridType } from '../common/grid.interface';
