+++ projects/igniteui-angular/src/lib/services/excel/excel-exporter-grid.spec.ts
@@ -1,6 +1,6 @@
-import { async, TestBed, fakeAsync, tick } from '@angular/core/testing';
+import { async, TestBed } from '@angular/core/testing';
 import { Component, ViewChild } from '@angular/core';
-import { IgxGridModule } from '../../grids/grid';
+import { IgxGridModule } from '../../grids/grid/public_api';
 import { IgxGridComponent } from '../../grids/grid/grid.component';
 import { IColumnExportingEventArgs, IRowExportingEventArgs } from '../exporter-common/base-export-service';
 import { ExportUtilities } from '../exporter-common/export-utilities';
