+++ projects/igniteui-angular/src/public_api.ts
@@ -69,7 +69,7 @@ export * from './lib/drop-down/public_api';
 export * from './lib/grids/hiding/hiding.module';
 export * from './lib/grids/common/enums';
 export * from './lib/grids/common/events';
-export * from './lib/grids/grid/index';
+export * from './lib/grids/grid/public_api';
 export * from './lib/grids/tree-grid/index';
 export * from './lib/grids/hierarchical-grid/index';
 export * from './lib/grids/columns/templates.directive';
